services/bridge-api/pyproject.toml
[project]
name = "sentinel-bridge-api"
version = "0.1.0"
description = "FastAPI bridge for Sentinel AI learning loop"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "pydantic==2.10.4",
  "requests==2.32.3",
  "python-dotenv==1.0.1"
]

[tool.uvicorn]
host = "127.0.0.1"
port = 8000
reload = true
